% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{asv_to_fasta}
\alias{asv_to_fasta}
\title{Export ASVs as fasta}
\usage{
asv_to_fasta(seqtab, path.out)
}
\arguments{
\item{seqtab}{matrix sequence table with samples as rows}

\item{path.out}{path where to save fasta}
}
\description{
Export ASVs as fasta
}
