% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV_tree_and_clusters.R
\name{ASV_tree_for_physeq}
\alias{ASV_tree_for_physeq}
\title{Build phylogenetic tree using NJ}
\usage{
ASV_tree_for_physeq(physeq, ncores = parallel::detectCores())
}
\arguments{
\item{physeq}{a phyloseq object}

\item{ncores}{number of cores to use}
}
\description{
Build phylogenetic tree using NJ
}
