% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{chimera_report}
\alias{chimera_report}
\title{Report chimera rate and reads}
\usage{
chimera_report(seqtab, seqtab.nochim)
}
\arguments{
\item{seqtab}{sequence table before chimera removal (samples as rows)}

\item{seqtab.nochim}{sequence table after chimera removal}
}
\description{
Report chimera rate and reads
}
