% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV_tree_and_clusters.R
\name{cluster_ASVs_physeq}
\alias{cluster_ASVs_physeq}
\title{Create ASV clusters based on phylogenetic distance}
\usage{
cluster_ASVs_physeq(physeq, hclust.method = "complete", threshold = 0.03)
}
\arguments{
\item{physeq}{a phyloseq object with tree data}

\item{hclust.method}{agglomeration method passed to `stats::hclust()`}

\item{threshold}{phylogenetic distance at which to cut the tree into clusters (0.03 = 97\% similarity)}
}
\description{
Adds a column to taxonomy, bridging the Genus-ASV gap for amplicons where the taxonomy is usually unresolved at Species level
Allows coarser-grain taxonomic analyses without losing unidentified ASVs
based on https://github.com/benjjneb/dada2/issues/947
}
