% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_tools.R
\name{compile_dist_pairs}
\alias{compile_dist_pairs}
\title{Pivot a dist object to a long dataframe}
\usage{
compile_dist_pairs(dist.mx, sample_subset = NULL)
}
\arguments{
\item{dist.mx}{distance matrix (dist object from vegan::)}

\item{sample_subset}{a vector of sample names to subset the distance matrix}
}
\description{
Possible to use only a subset of samples, provided their name
}
