% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_rates.R
\name{compute_classification_rates}
\alias{compute_classification_rates}
\title{Call the .compute_classrate() function over lists of ps objects if pertinent}
\usage{
compute_classification_rates(
  ps,
  ranks = c("Phylum", "Class", "Order", "Family", "Genus")
)
}
\arguments{
\item{ps}{a phyloseq object, or a named list of them (names become the `Dataset` column)}

\item{ranks}{taxonomic ranks to compute classification rates for}
}
\description{
Call the .compute_classrate() function over lists of ps objects if pertinent
}
