% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqDiversity.R
\name{div.fun}
\alias{div.fun}
\title{Compute Hill 0, 1 and 2, as well as Tail
Returns a list}
\usage{
div.fun(ps, idx)
}
\arguments{
\item{ps}{phyloseq object}

\item{idx}{a vector of index names}
}
\description{
Compute Hill 0, 1 and 2, as well as Tail
Returns a list
}
