% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diversities.R
\name{diversity_tibble}
\alias{diversity_tibble}
\title{Takes a phyloseq object as input and creates a tibble containing all the sample data and one column per index.}
\usage{
diversity_tibble(ps)
}
\arguments{
\item{ps}{a phyloseq object}
}
\description{
Takes a phyloseq object as input and creates a tibble containing all the sample data and one column per index.
}
