% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV_tree_and_clusters.R
\name{.check_genus_consistency}
\alias{.check_genus_consistency}
\title{Evaluates the consistency between ASV clusters and Genus-level classification.
Ideally, each cluster would have the same Genus-level taxonomy.
A cluster spanning multple genera suggests the clustering should be done at shorter phylogenetic distances (threshold)}
\usage{
.check_genus_consistency(tax_table)
}
\arguments{
\item{tax_table}{a taxonomy data.frame with ASVs as rownames and up to Genus taxonomy column, including a cluster_id column}
}
\description{
Evaluates the consistency between ASV clusters and Genus-level classification.
Ideally, each cluster would have the same Genus-level taxonomy.
A cluster spanning multple genera suggests the clustering should be done at shorter phylogenetic distances (threshold)
}
\keyword{internal}
