% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_rates.R
\name{.compute_classrate}
\alias{.compute_classrate}
\title{Compute classification rates}
\usage{
.compute_classrate(ps, Dataset = NA, ranks)
}
\arguments{
\item{ps}{a phyloseq object or a named list of phyloseq objects}

\item{Dataset}{String; creates a Dataset column in the output with that string as value. Managed automatically by compute_classification_rates() when named list of phylosq objects are provided.}

\item{ranks}{vector of taxonomic ranks to compute rate by}
}
\description{
Computes classification rates of amplicon taxonomy table using a
phyloseq object (or a list of them) as input. Produces
}
