% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV_tree_and_clusters.R
\name{.dendro_plot}
\alias{.dendro_plot}
\title{Plots the tree}
\usage{
.dendro_plot(hc, threshold)
}
\arguments{
\item{hc}{the output of hclust}

\item{threshold}{the phylogenetic distance threshold passed to}
}
\description{
Plots the tree
}
\keyword{internal}
