% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{drop_rare_asvs}
\alias{drop_rare_asvs}
\title{Remove ultra rare asvs}
\usage{
drop_rare_asvs(seqtab.nochim, at_least_n)
}
\arguments{
\item{seqtab.nochim}{sequence table (samples as rows)}

\item{at_least_n}{minimum total count for an ASV to be kept}
}
\description{
Remove samples with fewer than n sequences
as well as singletons, if any (shouldn't)
}
