% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{dropped_samples}
\alias{dropped_samples}
\title{Check if an expected file exists}
\usage{
dropped_samples(path_list)
}
\arguments{
\item{path_list}{character vector of expected output file paths}
}
\value{
`path_list` without the missing files
}
\description{
Filtering with minlen may yield empty samples (e.g. neg. controls);
list files that did survive filtering:
}
