% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqDiversity.R
\name{estimate_diversity}
\alias{estimate_diversity}
\title{Esitmate diversity (Shannon, Simpson, Tail)}
\usage{
estimate_diversity(ps, index = "Shannon")
}
\arguments{
\item{ps}{phyloseq object}

\item{index}{Richness, Shannon, Simpson or Tail}
}
\description{
Esitmate diversity (Shannon, Simpson, Tail)
}
