% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqDiversity.R
\name{estimate_Hill}
\alias{estimate_Hill}
\title{Functions to compute various diversity indices from phyloseq object}
\usage{
estimate_Hill(ps, q)
}
\arguments{
\item{ps}{phyloseq object}

\item{q}{Hill number order q (i-e 0, 1 or 2)}
}
\description{
Hill numbers
}
