% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{format_DECIPHER_for_dada2}
\alias{format_DECIPHER_for_dada2}
\title{format DECIPHER output for dada2}
\usage{
format_DECIPHER_for_dada2(ids, seqtab, ranks = NULL)
}
\arguments{
\item{ids}{output from  DECIPHER::IdTaxa}

\item{seqtab}{sequence table that was passed to DECIPHER::IdTaxa}

\item{ranks}{a vector describing the rank names to use, should start with 'rootrank'}
}
\description{
format DECIPHER output for dada2
}
