% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_viz.R
\name{gen_comm_plot_data}
\alias{gen_comm_plot_data}
\title{Community-plot data from a melted phyloseq object}
\usage{
gen_comm_plot_data(
  melted_ps,
  nTaxa = 20,
  taxRank = "Genus",
  grouping_vars,
  order_by = "alphabetical",
  seed = 1346134564
)
}
\arguments{
\item{melted_ps}{output of `phyloseq::psmelt()` or `psflashmelt()`}

\item{nTaxa}{number of taxa to show before lumping into "Others"}

\item{taxRank}{taxonomic rank to aggregate at}

\item{grouping_vars}{sample data columns to keep (e.g. for faceting)}

\item{order_by}{legend order: "alphabetical" or "abundance"}

\item{seed}{random seed for colour assignment}
}
\description{
long df prepared for barchart visualisation of community members
generates an "others" category for a specified number of taxa
at desired taxonomic rank
}
