% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{gen_qplots}
\alias{gen_qplots}
\title{Quality plots ----------------------------------------------------------
Generate qplots}
\usage{
gen_qplots(Fs, Rs, nsam = 4)
}
\arguments{
\item{Fs}{Forwards}

\item{Rs}{Reverse}

\item{nsam}{Number of plots to pick}
}
\description{
Quality plots ----------------------------------------------------------
Generate qplots
}
