% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{getN}
\alias{getN}
\title{Count reads retained in a dada-class or mergePairs object}
\usage{
getN(x)
}
\arguments{
\item{x}{a dada-class or mergePairs object}
}
\description{
Count reads retained in a dada-class or mergePairs object
}
