% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{load_checkpoint}
\alias{load_checkpoint}
\title{Reload an object saved with checkpoint()}
\usage{
load_checkpoint(name, dir)
}
\arguments{
\item{name}{short identifier used when checkpointing, e.g. "errF"}

\item{dir}{checkpoint directory}
}
\description{
Reload an object saved with checkpoint()
}
