% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgx.tools-package.R
\docType{package}
\name{mgx.tools-package}
\alias{mgx.tools}
\alias{mgx.tools-package}
\title{mgx.tools: Helper Functions for Metagenomics and Amplicon Sequencing Processing}
\description{
This package contains multiple helper functions for processing short reads shotgun and amplicon sequencing data when the objective is to go from raw data to phyloseq object. It includes rewrites of existing functions to allow them to use multiple cores when available, as well as short utility functions that help making informed decisions for data filtering and processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jorondo1/mgx.tools}
  \item Report bugs at \url{https://github.com/jorondo1/mgx.tools/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Rondeau-Leclaire \email{jonathan.rondeau.leclaire@gmail.com}

Authors:
\itemize{
  \item Jonathan Rondeau-Leclaire \email{jonathan.rondeau.leclaire@gmail.com}
}

}
\keyword{internal}
