% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{minimum_ASV_count}
\alias{minimum_ASV_count}
\title{Plot ASV counts with minimum sequence count}
\usage{
minimum_ASV_count(seqtab.nochim)
}
\arguments{
\item{seqtab.nochim}{sequence table (samples as rows)}
}
\description{
Plot ASV counts with minimum sequence count
}
