% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_rates.R
\name{plot_class_rates}
\alias{plot_class_rates}
\title{Plot classification rates}
\usage{
plot_class_rates(classrates)
}
\arguments{
\item{classrates}{output of `compute_classification_rates()`}
}
\description{
Plot classification rates
}
