% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{plot_track_change}
\alias{plot_track_change}
\title{Plot track changes}
\usage{
plot_track_change(track_change, reverse = FALSE)
}
\arguments{
\item{track_change}{output of `track_dada()` or `track_dada2()`}

\item{reverse}{whether reverse reads were denoised (paired-end)}
}
\description{
Plot track changes
}
