% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{primer_occurence}
\alias{primer_occurence}
\title{Compute primer occurence across all orientations}
\usage{
primer_occurence(fnFs, fnRs, FWD, REV)
}
\description{
Compute primer occurence across all orientations
}
