% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve_parallel.R
\name{rarecurve_to_df}
\alias{rarecurve_to_df}
\title{convert rarecurve output to data frame}
\usage{
rarecurve_to_df(rare_output, ps_object)
}
\arguments{
\item{rare_output}{rarefaction output}

\item{ps_object}{a phyloseq object}
}
\description{
convert rarecurve output to data frame
}
