% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_even_depth2.R
\name{rarefy_even_depth2}
\alias{rarefy_even_depth2}
\title{Parallelize phyloseq's rarefaction function with reproducible results}
\usage{
rarefy_even_depth2(
  physeq,
  sample.size = min(phyloseq::sample_sums(physeq)),
  rngseed = FALSE,
  replace = TRUE,
  trimOTUs = TRUE,
  verbose = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{sample.size}{depth to rarefy to; samples below it are dropped}

\item{rngseed}{numeric seed for reproducible subsampling, or FALSE}

\item{replace}{subsample with replacement}

\item{trimOTUs}{remove taxa with zero counts after rarefying}

\item{verbose}{print messages}

\item{ncores}{number of cores to use}
}
\description{
Parallelize phyloseq's rarefaction function with reproducible results
}
