% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{remove_ultra_rare}
\alias{remove_ultra_rare}
\title{Remove samples with fewer than n sequences once taxa removed}
\usage{
remove_ultra_rare(seqtab, taxonomy, n)
}
\arguments{
\item{seqtab}{matrix sequence table with samples as rows}

\item{taxonomy}{taxonomic table with unique ID as rows (e.g. OTU, ASV, Species)}

\item{n}{samples with fewer than n sequences are removed}
}
\description{
Remove samples with fewer than n sequences once taxa removed
}
