% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{samdat_as_tibble}
\alias{samdat_as_tibble}
\title{1. Get a phyloseq object's sample data as tibble
Creates a Sample column with the sample names}
\usage{
samdat_as_tibble(ps, strings_as_factors = TRUE)
}
\description{
1. Get a phyloseq object's sample data as tibble
Creates a Sample column with the sample names
}
