% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{save_qplots}
\alias{save_qplots}
\title{Export qplot}
\usage{
save_qplots(plot_list, out_dir, step_id)
}
\arguments{
\item{plot_list}{output of gen_qplot}

\item{out_dir}{save destination}

\item{step_id}{identifier for the step (e.g. "raw_samples")}
}
\description{
Export qplot
}
