% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity.R
\name{sparsity}
\alias{sparsity}
\title{Compute sparsity from a matrix where samples are column}
\usage{
sparsity(input)
}
\arguments{
\item{input}{matrix, tibble, data.frame or phyloseq object}
}
\description{
Compute sparsity from a matrix where samples are column
}
