% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{subset_samples}
\alias{subset_samples}
\title{utility tools to work with phyloseq objects
1. samdat_as_tibble()
2.}
\usage{
subset_samples(seqtab, samples)
}
\arguments{
\item{seqtab}{matrix sequence table with samples as rows}

\item{samples}{vector of sample names}
}
\description{
Keep samples with metadata
}
