% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_tools.R
\name{vst_ps_to_mx}
\alias{vst_ps_to_mx}
\title{Distance-based analyses
Variance-stabilizing transformation}
\usage{
vst_ps_to_mx(ps)
}
\arguments{
\item{ps}{phyloseq object}
}
\description{
Distance-based analyses
Variance-stabilizing transformation
}
