% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblePhyloseq.R
\name{assemble_phyloseq}
\alias{assemble_phyloseq}
\title{Assemble phyloseq object from a table that contains both taxonomy and sample identifiers in the columns
Tailored to work with the tibble-format output of parse_*() functions (mgx.tools package under development)}
\usage{
assemble_phyloseq(
  abunTable,
  sampleData,
  min_sample_size = 100,
  min_taxa_count = 1,
  unique_taxa_ID = "Species",
  filtering = FALSE,
  justBacteria = FALSE,
  onlySpecies = FALSE,
  seqdepth_plot = TRUE
)
}
\arguments{
\item{sampleData}{data frame with sample data and matching sample IDs as row names}

\item{min_sample_size}{default 100; samples with fewer sequences are discarded}

\item{min_taxa_count}{default 1; taxa with fewer sequences overall (sum of all samples) are removed}

\item{unique_taxa_ID}{default "Species"; Column name containing unique taxa IDs. For shotgun (Metaphlan, Kraken, etc.), usually "Species".}

\item{filtering}{default FALSE; deprecated}

\item{justBacteria}{default FALSE; filter to keep only bacteria counts}

\item{abundTable}{wide tibble with taxonomic ranks and sample IDs as column; one species (or ASV) per row}
}
\description{
Assemble phyloseq object from a table that contains both taxonomy and sample identifiers in the columns
Tailored to work with the tibble-format output of parse_*() functions (mgx.tools package under development)
}
