% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_tools.R
\name{compute_pcoa}
\alias{compute_pcoa}
\title{PCOA
Return a list of 3 elements :
1. the phyloseq sample_data with 2 first PCo added
2. the eigenvalues
3. the distance/dissimilarity matrix}
\usage{
compute_pcoa(ps, dist, all_coordinates = FALSE, vst = FALSE)
}
\arguments{
\item{ps}{phyloseq object}

\item{dist}{distance name (choose from vegan)}

\item{all_coordinates}{default FALSE; whether to include a table with all coordinates instead of just adding the first 2 to the sample_data table}

\item{vst}{default FALSE, whether to use variance-stabilizing transformation}
}
\description{
PCOA
Return a list of 3 elements :
1. the phyloseq sample_data with 2 first PCo added
2. the eigenvalues
3. the distance/dissimilarity matrix
}
