% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{.allOrients}
\alias{.allOrients}
\title{Find all orientation of primers in the data
stolen from DADA2 tutorial https://benjjneb.github.io/dada2/tutorial.html}
\usage{
.allOrients(primer)
}
\description{
Find all orientation of primers in the data
stolen from DADA2 tutorial https://benjjneb.github.io/dada2/tutorial.html
}
\keyword{internal}
