% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_viz.R
\name{gen_comm_plot_data}
\alias{gen_comm_plot_data}
\title{long df prepared for barchart visualisation of community members
generates an "others" category for a specified number of taxa
at desired taxonomic rank}
\usage{
gen_comm_plot_data(
  melted_ps,
  nTaxa = 20,
  taxRank = "Genus",
  grouping_vars,
  order_by = "alphabetical",
  seed = 1346134564
)
}
\description{
long df prepared for barchart visualisation of community members
generates an "others" category for a specified number of taxa
at desired taxonomic rank
}
