% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve_parallel.R
\name{rarecurve_parallel}
\alias{rarecurve_parallel}
\title{Paralellized version of vegan::rarecurve()
Shamelessly adapted by Gemini}
\usage{
rarecurve_parallel(
  x,
  step = 1,
  sample,
  xlab = "Sample Size",
  ylab = "Species",
  label = TRUE,
  col,
  lty,
  tidy = FALSE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{x}{count table with sample as rownames and taxa as colnames}
}
\description{
Paralellized version of vegan::rarecurve()
Shamelessly adapted by Gemini
}
