% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_even_depth2.R
\name{rarefy_even_depth2}
\alias{rarefy_even_depth2}
\title{Paralellize phyloseq's rarefaction function
Shamelessly stolen and fed to Gemini}
\usage{
rarefy_even_depth2(
  physeq,
  sample.size = min(phyloseq::sample_sums(physeq)),
  rngseed = FALSE,
  replace = TRUE,
  trimOTUs = TRUE,
  verbose = FALSE,
  ncores = detectCores() - 1
)
}
\description{
Paralellize phyloseq's rarefaction function
Shamelessly stolen and fed to Gemini
}
