% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_even_depth2.R
\name{rarefy_even_depth2}
\alias{rarefy_even_depth2}
\title{Parallelize phyloseq's rarefaction function with reproducible results}
\usage{
rarefy_even_depth2(
  physeq,
  sample.size = min(phyloseq::sample_sums(physeq)),
  rngseed = FALSE,
  replace = TRUE,
  trimOTUs = TRUE,
  verbose = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\description{
Parallelize phyloseq's rarefaction function with reproducible results
}
