% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurve.R
\name{rcurve}
\alias{rcurve}
\title{stolen from https://github.com/Russel88/MicEco/blob/master/R/rcurve.R
minor adjustments to transform otu as matrix}
\usage{
rcurve(physeq, subsamp = 10^c(1:5), trim = TRUE, add_sample_data = TRUE)
}
\description{
stolen from https://github.com/Russel88/MicEco/blob/master/R/rcurve.R
minor adjustments to transform otu as matrix
}
