% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{run_cutadapt}
\alias{run_cutadapt}
\title{CUTADAPT wrapper function}
\usage{
run_cutadapt(i, cutadapt_path, R1.flags, R2.flags)
}
\description{
Adapted cutadapt in function to allow using it within mclapply
}
