% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{samdat_as_tibble}
\alias{samdat_as_tibble}
\title{Get a phyloseq object's sample data as tibble}
\usage{
samdat_as_tibble(ps, strings_as_factors = TRUE)
}
\arguments{
\item{ps}{a phyloseq object}

\item{strings_as_factors}{convert strings to factors}
}
\description{
Creates a Sample column with the sample names
}
