% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{subset_tax_table}
\alias{subset_tax_table}
\title{Filter taxonomy}
\usage{
subset_tax_table(taxonomy, seqtab, min_seq)
}
\arguments{
\item{taxonomy}{taxonomic table with unique ID as rows (e.g. OTU, ASV, Species)}

\item{seqtab}{matrix sequence table with samples as rows}

\item{min_seq}{taxa with fewer than min_seq overall (sum of all samples) are removed}
}
\description{
ASVs classified at the kingdom level and present in seqtab
}
