% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_glom2.R
\name{tax_glom2}
\alias{tax_glom2}
\title{Rewritten `tax_glom()`}
\usage{
tax_glom2(ps, taxrank)
}
\arguments{
\item{ps}{phyloseq object}

\item{taxrank}{name of rank to agglomerate to}
}
\description{
This new version of tax_glom updates the index names to the
agglomerated taxrank, to avoid e.g. keeping an arbitrary ASV for
a genus that encompasses multiple ASVs post agglomeration.
Finally, in the event that some species have missing taxrank,
those are simply removed entirely.
Further detail :
I created what I think is a better version of this function,
which I creatively called tax_glom2(). To explain why, consider three things:
1. tax_glom() will always keep the  original OTU names (taxa_names) regardless
of the agglomeration. This means, if you have 20 OTU in the same Genus, you
end up with one Genus (expected), but its taxa_name will be one of the OTUs
(probably the first one the algorithm encounters). With shotgun data that's
especially annoying because the phyloseq OTU value (the taxa_name attribute)
will be a species. These are the unique keys that link your otu_table to your
tax_table. Then, when you use software that uses these (e.g. for differential
abundance analysis), they output the name of a species when, in fact, you have
a genus-agglomerated phyloseq object. So my first enhancement is for
tax_glom2() to generate a phyloseq object with the taxa_names updated to
reflect the taxrank you agglomerate to.
2. with more and more MAGs being added to public databases, there are
instances where an organism will have a species-level label, but no genus or
family. An example : Clostridia bacterium UC5.1-1E11 only has a class-level
classification; thus Order, Family, and Genus will all be NA. So keeping my
previous modification to tax_glom in mind, agglomeration at the genus level
will pool ALL read counts of NA genera together, regardless of their higher
taxonomy. Since these genomes have not been classified at these levels, all
one can say is that it's a "species" of the class Clostridia, probably because
marker genes from that clade were found in that genome; and that it is different
enough from all other known clostridia genomes so that we know it's an unknown
species of Clostridia. In this case, if you are interested in genus-level
metagenomics, this genome is as good as an undetected one, so tax_glom2()
simply removes these counts when you agglomerate to Genus level.
3. Agglomerating creates a new taxa which encompasses all the counts
of all taxa that are NA or 'Unclassified' ; therefore, they are
removed post-agglomeration and counts go down.
Comments mostly and shamelessly generated by ChatGPT 4o
Function to melt, aggregate, and rebuild a phyloseq object
}
