% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_viz.R
\name{topTaxa}
\alias{topTaxa}
\title{find top N taxa and generate a tibble of taxa with Others category}
\usage{
topTaxa(psmelt, taxLvl, topN)
}
\description{
find top N taxa and generate a tibble of taxa with Others category
}
