% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_utils.R
\name{track_dada}
\alias{track_dada}
\title{Read tracker
compile reads across samples in long format
also computes changes between steps (last column flags this since format is long)}
\usage{
track_dada(
  out.N,
  out,
  sample.names,
  dadaFs,
  dadaRs = NULL,
  mergers_pooled = mergers_pooled,
  seqtab.nochim = seqtab.nochim
)
}
\description{
Read tracker
compile reads across samples in long format
also computes changes between steps (last column flags this since format is long)
}
