% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeqUtils.R
\name{viz_seqdepth}
\alias{viz_seqdepth}
\title{Visualise sequence count distribution across samples}
\usage{
viz_seqdepth(seqtab, breaks = 100, log_count = FALSE)
}
\arguments{
\item{seqtab}{matrix sequence table with samples as rows}
}
\description{
Visualise sequence count distribution across samples
}
